upload uml
Note that the uml has not yet been completed and there are several flaws with it. We are not sure about implementing certain design patterns; for example facade and iterator pattern. The javaFX view/controller part could probably have a facade to it instead of implementing so many interfaces (for example IChatController) since these classes are dependent on eachother anyways. For example you  cannot switch out a single JavaFX component for something else, but you could switch it out to another JavaFX component - is there any use in this?